Have separate directories for default and user-defined channels
Resolves #1
