+++ test/MessageBus.Specs/DefaultChannel/PublishMethod/Validation.cs
@@ -3,7 +3,7 @@ using FluentAssertions;
 using Machine.Specifications;
 using MessageBus.Specs.Context;
 
-namespace MessageBus.Specs.PublishMethod;
+namespace MessageBus.Specs.DefaultChannel.PublishMethod;
 
 [Subject("Publishing: Validation")]
 class publishing_validation_context : bus_context
@@ -13,7 +13,9 @@ class publishing_validation_context : bus_context
 [Tags(tag.validation)]
 class given_null_message : publishing_validation_context
 {
-    Because of = () => publishing = () => bus.Publish(null);
+    Because of =
+        // ReSharper disable once AssignNullToNotNullAttribute
+        () => publishing = () => bus.Publish(null);
 
     It should_throw_argument_null_exception = () => publishing.Should().Throw<ArgumentNullException>();
